refactor: getModal util 함수 추가
